Adding Readme, Amiga Format ADFs, Ariticles plus unpacked source code from each of the coverdisks
Initial commit to provide the pristine version of the source code for Menace as it was provided by Amiga Format in the 1990 editions of the magazine.
Adding PDF's containing just the coverpage and the Dave Jones articles from Amiga Format and the ADF files (that were downloaded from Amiga Magazine Rack but the files have currently gone missing from that site, the links are down).