Changes to upgrade the system takeover to work on A1200 NDK3.2
These changes are ones suggested from the safestart.i from the Bare Metal Amiga Programming Book.
These currently are working on the A1200 setup in WinUAE (A1200 with 4Mb of Fast ram).
The game will start and stop but has graphical glitches and missile collision detection is not working.
Removing the assembled Menace Binary and the .gs pre-compiled headers that DevPac has made