blitter wait after missile collision check
The missile to alien collision if performed by a blitter test. There was no wait to see if the blitter finished before reading the result of the blit. This code appears to work fine on a stock A500, but on an A1200 (wtih fast ram), the CPU can check the result of the blit before it has finished. The busy wait here is wasting CPU cycles until the blit has finished.