C# script to convert "Aliens" into PNG files
The C# script can read in the "Aliens" bitplane graphics  from Meance plus the palette data (which is stored in a text format to be compiled into the program).  It outputs an RGB .png for each of the aliens stored in the file.
The converted .png files have been added as part of this push.