Adding script to read the backgrounds into a PNG
Reading the backgrounds from Menace.s into a PNG file. Currently the palette has not been read in so have just converted the color index into an RGB value so it's possible to see that the pixel data is correct (but the colours are not).
There appear to be only 12 (16px x 16px) blocks in the background, but the code implies that the level should have 16