Reading menace backgroundtable and converting to xml
The menace background table defines how to draw the background tiles. They are stored in the code in 4bit per tile format. Converting that code into XML so the background data could be loaded into an editor such as "Tiled".
The C# has output an XML file but it has not yet been tested.