The loading of the asset file into chip memory will quit if the read would overwrite then end of the buffer allocated
Currently there is a 50Kb allocated chip memory buffer to read the file into before system takeover. If the file that is read in could possibly overwrite the end of the 50Kb buffer then the file is closed and all memory is freed and the application is closed.
This will be more useful later on as we want to load in many files into the buffer (not just one).