incomplete code for loading a list of strings from an asset file
This code isn't finished. I was going to read 1 line at a time from the file into a single reused buffer, but the FGets method doesn't appear to be available in DevPac. Will probably have to load into a buffer myself and decode the string from there.

The asset file I have made currently only contains 1 line and to start with this is where we will get the name of the .mod file asset to be loaded (instead of hard coding the name into the source)