Allocate fast RAM for file / Read in the asset list number as BCD (up to 4 digits)
Allocating some fast memory for files (50Kb at the moment).
When reading the number for the asset list, treat it as binary coded decimal and read the digits into a register (instead of the stack), limits the number to a max of 4 digits (2 digits will be enough for this purpose anyway).