Read asset list Chip/Fast character, find start of the filename string
During the asset list read, this now knows the number assets to load an whether to put them in chip or fast. The start address of the string for the filename is know, but the string is not zero delimited. The end of the string newline is found but currently the data that was read in isn't used to try and load the asset (the string would need to be converted to 0 delimited before it could be passed to the load file OS routine).