Loading the asset list and loading the files from name that are read from that list / Allocating fast RAM for assets.
Allocating some fast ram and creating a loader for the list of assets to load into either chip of fast.
currently each asset is loaded but the pointer to it and it's size is not stored correctly (it's always just overwriting the levelmusic pointer), and the asset list is currently loading the same .mod file twice for testing purposes.

The asset file list is
number,RAM,filename

where the number is going to be the number of versions of that asset to load (not currently implemented).
RAM is etiher C or F to indicate the type of RAM to store the asset in.
filename is the name of the file to load and must end in the single newline character

any line with a newline character at the start indicate the end fo the file and reading more lines is stopped.