the C# script is able to decode the ships from ships.s back into PNG files (with some error)
Can parse the data into ships.txt and read them into .png files.
As these are sprites, the data for the ship is split into front /back. Back is 44 pixels high, front is 22 pixels high (back is higher to accommodate outriders).
Appears to be a vertical line corrupted on the last line of the back image of each sprite, not sure why at this point.